+++ class/mod.lua
@@ -20,3 +20,7 @@ Mod.find = function(mod_id)
 
     return nil
 end
+
+
+
+return Mod