+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.1.21
 * Damager : Added `add_offset`.
+* Mod : Class reference is now public (whoops again).
 
 ### v1.1.20
 * Helper : Fixed specific error in `log_struct`.